tenant_server/pyproject.toml
[project]
name = "tenant-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "sentencepiece>=0.2.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.3",
    "uvicorn>=0.34.2",
]
